Arithmetic series helper
Just a small bit of code that is supposed to paired with a form to solve
arithmetic series questions.

Pretty sloppy but wanted to get a working version out as this was made
at the request of a few friends but I will work on making the code more
efficient and cleaner to read in the future.